requirements.txt
openai
googletrans==4.0.0-rc1
groq
transformers
torch
Pillow
PyAudio
streamlit_webrtc
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
gtts 
playsound==1.2.2 
keyboard
streamlit 
pyttsx3 
pydub
sounddevice 
numpy 
#skin analysis  
timm 
accelerate
google-generativeai
torchvision 
torchaudio  
scipy 
librosa 
noisereduce 
joblib 
